backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Isuru Jinasena <iujinasena@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11,<3.12"
fastapi = "^0.109.1"
python-dotenv = "^1.0.0"
python-multipart = "0.0.9"
uvicorn = "^0.23.2"
PyMuPDF = "1.24.9"
ollama = "0.3.3"
plum-dispatch = "1.7.4"
pillow = "^11.0.0"
torch = "^2.5.1"
transformers = "^4.46.3"
timm = "^1.0.11"
azure-ai-documentintelligence = "^1.0.0b4"
azure-ai-formrecognizer = "^3.3.3"
openai = "^1.55.1"
spire-doc = "12.7.1"
beautifulsoup4 = "^4.12.3"
markdownify = "^0.14.1"

[build-system]
requires = [ "poetry-core" ]
build-backend = "poetry.core.masonry.api"
